pyproject.toml
[project]
name = "agent-builder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.6"
dependencies = [
    "django>=5.2.6",
    "django-extensions>=4.1",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "django-cors-headers>=4.7.0",
    "djangorestframework-simplejwt>=5.5.0",
    "dq-db-manager",
]

[tool.uv.sources]
dq-db-manager = { git = "https://github.com/Data-Quotient/dq_db_manager.git" }
